Tournaments/requirements.txt
django
pydantic-settings
gunicorn
Pillow
django-allauth
requests
pyjwt
cryptography
